packages/sdk-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "veto-sdk"
version = "0.1.0"
description = "A guardrail system for AI agent tool calls - Python SDK"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
authors = [
    { name = "Plaw, Inc." }
]
keywords = [
    "ai",
    "agent",
    "guardrail",
    "tool-calls",
    "safety",
    "llm",
    "openai",
    "anthropic",
    "langchain"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "pyyaml>=6.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
openai = [
    "openai>=1.0",
]
anthropic = [
    "anthropic>=0.18",
]
langchain = [
    "langchain>=0.1",
]

[project.urls]
Homepage = "https://veto.run"
Repository = "https://github.com/VulnZap/veto"
Documentation = "https://github.com/VulnZap/veto/tree/master/packages/sdk-python"

[tool.hatch.build.targets.wheel]
packages = ["src/veto_sdk"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
